% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_maker.R
\name{comparison_maker}
\alias{comparison_maker}
\title{Comparison Maker}
\usage{
comparison_maker(data, x, comparisons = "all")
}
\arguments{
\item{data}{The dataset you are inputting}

\item{x}{The column name that contains the groups you are wanting to compare}

\item{comparisons}{Default is "all". If you do not want to compare all variables to each other within your grouping column, then you can specify which comparisons you want instead.}
}
\value{
comparisons
}
\description{
Comparison Maker
}
\examples{
comparison_maker(ToothGrowth, supp)
}
