% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_data.R
\name{prep_data}
\alias{prep_data}
\title{Title}
\usage{
prep_data(data, x, y, groups = NULL)
}
\arguments{
\item{data}{data}

\item{x}{x}

\item{y}{y}

\item{groups}{groups}
}
\value{
data
}
\description{
Title
}
\examples{
x
}
