% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comparison_maker.R
\name{comparison_maker}
\alias{comparison_maker}
\title{Comparison Maker}
\usage{
comparison_maker(data, x, comparisons = "all")
}
\arguments{
\item{data}{The dataset you are inputting}

\item{comparisons}{Default is "all". If you do not want to compare all variables to each other within your grouping column, then you can specify which comparisons you want instead.}

\item{groups}{The column name that contains the groups you are wanting to compare}
}
\value{
A list of comparisons to be made
}
\description{
Comparison Maker
}
\examples{
input <- comparison_maker(ds, "Sample")
plot(ds, aes(x=Sample, y=Values))+
 geom_boxplot()+
 stat_compare_means(comparisons = input)
}
