% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_stats.R
\name{run_stats}
\alias{run_stats}
\title{Title}
\usage{
run_stats(
  data,
  x,
  y,
  groups = NULL,
  test = "ANOVA",
  interactions = T,
  parametric = T
)
}
\arguments{
\item{data}{dataset}

\item{x}{x}

\item{y}{y}

\item{groups}{groups}

\item{test}{test}

\item{interactions}{ints}

\item{parametric}{parametric}
}
\value{
out
}
\description{
Title
}
\examples{
x
}
